drivers/perf: Add T-HEAD c900_pmu support
T-HEAD c900_pmu follows riscv pmu spec and extend bunches of CPU
hardware events. The 'perf stat' and 'perf record' all support the
hardware events. There are 32 counters in the pmu. The counter0-2
are fix mapped to cycles, timer, instructions and the rest of 28
counters could be dynamically mapped for any events.

Signed-off-by: Guo Ren <guoren@linux.alibaba.com>